Everywhere: Support sorting based on senders instead of recipients
This patch rewrites most of the internal logic to be able to handle
senders instead of just recipients on a per rule basis.

Rules now get evaluated in order and applied to all recipients or all
senders of a message depending on the rule's settings. If any match
is found, the matching algorithm returns and the resulting slug gets
used for sorting the message. If no match is found, the next rule is
tried.

Alongside this new UI has been added to make debugging rules in the
settings a little easier.

Resolves #54